fix(zcncore): added SmartContractExecutor
